+++ crates/primitives/convert/src/felt.rs
@@ -36,8 +36,8 @@ mod tests {
         assert_eq!(felt_to_u64(&Felt::THREE).unwrap(), 3);
         assert_eq!(felt_to_u64(&Felt::from(u32::MAX)).unwrap(), u32::MAX as u64);
         assert_eq!(felt_to_u64(&Felt::from(u64::MAX)).unwrap(), u64::MAX);
-        assert!(felt_to_u64(&(Felt::from(u64::MAX) + Felt::ONE)).is_err());
-        assert!(felt_to_u64(&Felt::MAX).is_err());
+        assert!(matches!(felt_to_u64(&(Felt::from(u64::MAX) + Felt::ONE)), Err(FeltToU64Error)));
+        assert!(matches!(felt_to_u64(&Felt::MAX), Err(FeltToU64Error)));
     }
 
     #[test]
@@ -48,7 +48,7 @@ mod tests {
         assert_eq!(felt_to_u128(&Felt::THREE).unwrap(), 3);
         assert_eq!(felt_to_u128(&Felt::from(u64::MAX)).unwrap(), u64::MAX as u128);
         assert_eq!(felt_to_u128(&Felt::from(u128::MAX)).unwrap(), u128::MAX);
-        assert!(felt_to_u128(&(Felt::from(u128::MAX) + Felt::ONE)).is_err());
-        assert!(felt_to_u128(&Felt::MAX).is_err());
+        assert!(matches!(felt_to_u128(&(Felt::from(u128::MAX) + Felt::ONE)), Err(FeltToU128Error)));
+        assert!(matches!(felt_to_u128(&Felt::MAX), Err(FeltToU128Error)));
     }
 }