+++ crates/primitives/convert/src/lib.rs
@@ -1,7 +1,5 @@
 mod felt;
-mod state_update;
 mod to_felt;
 
 pub use felt::{felt_to_u128, felt_to_u64};
-pub use state_update::ToStateUpdateCore;
 pub use to_felt::ToFelt;