+++ crates/primitives/state_update/src/into_starknet_core.rs
@@ -170,3 +170,38 @@ impl From<NonceUpdate> for starknet_core::types::NonceUpdate {
         Self { contract_address: nonce_update.contract_address, nonce: nonce_update.nonce }
     }
 }
+
+#[cfg(test)]
+mod test {
+    use starknet_types_core::felt::Felt;
+
+    use crate::tests::dummy_state_diff;
+
+    use super::*;
+
+    #[test]
+    fn test_state_update_core_convertion() {
+        let state_update = StateUpdate {
+            block_hash: Felt::from_hex_unchecked("0x1234"),
+            old_root: Felt::from_hex_unchecked("0x5678"),
+            new_root: Felt::from_hex_unchecked("0x9abc"),
+            state_diff: dummy_state_diff(),
+        };
+
+        let core_state_update: starknet_core::types::StateUpdate = state_update.clone().into();
+        let converted_state_update: StateUpdate = core_state_update.into();
+
+        assert_eq!(state_update, converted_state_update);
+    }
+
+    #[test]
+    fn test_pending_state_update_core_convertion() {
+        let pending_state_update =
+            PendingStateUpdate { old_root: Felt::from_hex_unchecked("0x5678"), state_diff: dummy_state_diff() };
+
+        let core_pending_state_update: starknet_core::types::PendingStateUpdate = pending_state_update.clone().into();
+        let converted_pending_state_update: PendingStateUpdate = core_pending_state_update.into();
+
+        assert_eq!(pending_state_update, converted_pending_state_update);
+    }
+}