+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- test: add unitests primitives
 - fix: #225
 - feat: experimental block production and mempool
 - refactor: L1BlockMetric is intialized inside the EthereumClient new function